refactor to separate router and business logic
