+++ server/api/routers/inference/__init__.py
@@ -1,1 +1,3 @@
 from .inference import router
+
+__all__ = ["router"]
