+++ server/api/routers/inference/inference.py
@@ -1,17 +1,13 @@
 import uuid
-from typing import Dict, Optional, List, Any, Tuple
+from typing import Optional
 from fastapi import APIRouter, HTTPException, Header
-from pydantic import BaseModel
-import json
-import re
-from dataclasses import dataclass
-from enum import Enum
+from atomic_agents import BasicChatInputSchema
 
-# Correct v2.0 imports
-from atomic_agents import AtomicAgent, AgentConfig, BasicChatInputSchema, BasicChatOutputSchema
-from atomic_agents.context import ChatHistory, SystemPromptGenerator
-import instructor
-from openai import OpenAI
+# Import business logic from separate modules
+from .models import ChatRequest, ChatResponse, IntegrationType, ProjectAction
+from .services import ChatProcessor, AgentSessionManager, ToolExecutor, ResponseFormatter
+from .analyzers import MessageAnalyzer, ResponseAnalyzer
+from .factories import ModelManager
 from core.config import settings
 
 try:
@@ -26,390 +22,6 @@ router = APIRouter(
     prefix="/inference",
     tags=["inference"],
 )
-
-class ChatRequest(BaseModel):
-    message: str
-
-class ChatResponse(BaseModel):
-    message: str
-    session_id: str
-    tool_results: Optional[List[Dict]] = None
-
-class IntegrationType(Enum):
-    NATIVE = "native_atomic_agents"
-    MANUAL = "manual_execution"
-    MANUAL_FAILED = "manual_execution_failed"
-
-class ProjectAction(Enum):
-    LIST = "list"
-    CREATE = "create"
-
-@dataclass
-class ToolResult:
-    success: bool
-    action: str
-    namespace: str
-    message: str = ""
-    result: Any = None
-    integration_type: IntegrationType = IntegrationType.MANUAL
-
-@dataclass
-class ModelCapabilities:
-    supports_tools: bool
-    instructor_mode: instructor.Mode
-    
-# Store agent instances to maintain conversation context
-agent_sessions: Dict[str, AtomicAgent] = {}
-
-# Constants
-TOOL_CALLING_MODELS = [
-    "llama3.1", "llama3.1:8b", "llama3.1:70b", "llama3.1:405b",
-    "mistral-nemo", "firefunction-v2", "hermes3"
-]
-
-TEMPLATE_INDICATORS = [
-    "[number of projects]", "[project list]", "[namespace]", "[projects]",
-    "{{", "}}", "${", "[total]", "[count]", "**[list of projects",
-    "**[projects", "[list of projects", "I will use the project tool",
-    "To view the list, I will", "**[namespace", "currently **[",
-]
-
-PROJECT_KEYWORDS = ["project", "list", "create", "show", "namespace"]
-
-NAMESPACE_PATTERNS = [
-    r"in\s+(\w+)\s+namespace",
-    r"namespace\s+(\w+)",
-    r"in\s+(\w+)(?:\s|$)",
-    r"from\s+(\w+)(?:\s|$)",
-    r"(\w+)\s+namespace"
-]
-
-PROJECT_ID_PATTERNS = [
-    r"create\s+(?:project\s+)?(?:called\s+)?['\"]?(\w+)['\"]?",
-    r"new\s+project\s+['\"]?(\w+)['\"]?",
-    r"project\s+['\"]?(\w+)['\"]?"
-]
-
-CREATE_KEYWORDS = ["create", "new", "add", "make"]
-EXCLUDED_NAMESPACES = ["the", "a", "an", "my", "projects", "project"]
-
-class MessageAnalyzer:
-    """Handles message analysis and parameter extraction"""
-    
-    @staticmethod
-    def extract_namespace(message: str) -> str:
-        """Extract namespace from user message or return default"""
-        message_lower = message.lower()
-        
-        # Look for explicit namespace mentions
-        for pattern in NAMESPACE_PATTERNS:
-            match = re.search(pattern, message_lower)
-            if match:
-                namespace = match.group(1)
-                if namespace not in EXCLUDED_NAMESPACES:
-                    return namespace
-        
-        return "test"  # default
-
-    @staticmethod
-    def extract_project_id(message: str) -> Optional[str]:
-        """Extract project ID from create project messages"""
-        message_lower = message.lower()
-        
-        for pattern in PROJECT_ID_PATTERNS:
-            match = re.search(pattern, message_lower)
-            if match:
-                return match.group(1)
-        
-        return None
-
-    @staticmethod
-    def determine_action(message: str) -> ProjectAction:
-        """Determine if user wants to create or list projects"""
-        message_lower = message.lower()
-        return ProjectAction.CREATE if any(word in message_lower for word in CREATE_KEYWORDS) else ProjectAction.LIST
-
-    @staticmethod
-    def is_project_related(message: str) -> bool:
-        """Check if message is project-related"""
-        return any(word in message.lower() for word in PROJECT_KEYWORDS)
-
-class ResponseAnalyzer:
-    """Handles response analysis and validation"""
-    
-    @staticmethod
-    def is_template_response(response: str) -> bool:
-        """Detect if response contains template placeholders"""
-        response_lower = response.lower()
-        return any(indicator.lower() in response_lower for indicator in TEMPLATE_INDICATORS)
-
-    @staticmethod
-    def needs_manual_execution(response: str, message: str) -> bool:
-        """Determine if manual tool execution is needed"""
-        if not MessageAnalyzer.is_project_related(message):
-            return False
-            
-        return (
-            ResponseAnalyzer.is_template_response(response) or
-            len(response) < 50 or
-            "I don't have access" in response or
-            "cannot directly" in response or
-            "[list of projects" in response or
-            "**[" in response or
-            "I will use the project tool" in response
-        )
-
-class ModelManager:
-    """Handles model capabilities and configuration"""
-    
-    @staticmethod
-    def get_capabilities(model_name: str) -> ModelCapabilities:
-        """Get model capabilities based on model name"""
-        model_lower = model_name.lower()
-        supports_tools = any(supported in model_lower for supported in TOOL_CALLING_MODELS)
-        
-        return ModelCapabilities(
-            supports_tools=supports_tools,
-            instructor_mode=instructor.Mode.TOOLS if supports_tools else instructor.Mode.JSON
-        )
-
-    @staticmethod
-    def create_client(capabilities: ModelCapabilities) -> Any:
-        """Create instructor client with appropriate mode"""
-        ollama_client = OpenAI(
-            base_url=settings.ollama_host,
-            api_key=settings.ollama_api_key,
-        )
-        
-        return instructor.from_openai(ollama_client, mode=capabilities.instructor_mode)
-
-class ToolExecutor:
-    """Handles tool execution (both native and manual)"""
-    
-    @staticmethod
-    def execute_manual(message: str) -> ToolResult:
-        """Manually execute tool based on message analysis"""
-        try:
-            projects_tool = ProjectsTool()
-            action = MessageAnalyzer.determine_action(message)
-            namespace = MessageAnalyzer.extract_namespace(message)
-            
-            if action == ProjectAction.CREATE:
-                project_id = MessageAnalyzer.extract_project_id(message)
-                if not project_id:
-                    return ToolResult(
-                        success=False,
-                        action=action.value,
-                        namespace=namespace,
-                        message="Please specify a project name to create. For example: 'Create project my_app'"
-                    )
-                
-                tool_input = ProjectsToolInput(
-                    action=action.value,
-                    namespace=namespace,
-                    project_id=project_id
-                )
-            else:
-                tool_input = ProjectsToolInput(action=action.value, namespace=namespace)
-            
-            print(f"🔧 [Manual Tool] Executing {action.value} in namespace '{namespace}'" + 
-                  (f" with project_id '{tool_input.project_id}'" if hasattr(tool_input, 'project_id') and tool_input.project_id else ""))
-            
-            result = projects_tool.run(tool_input)
-            
-            return ToolResult(
-                success=result.success,
-                action=action.value,
-                namespace=namespace,
-                result=result,
-                integration_type=IntegrationType.MANUAL
-            )
-            
-        except Exception as e:
-            print(f"❌ [Manual Tool] Error: {str(e)}")
-            return ToolResult(
-                success=False,
-                action="unknown",
-                namespace="unknown",
-                message=f"Tool execution failed: {str(e)}",
-                integration_type=IntegrationType.MANUAL_FAILED
-            )
-
-class ResponseFormatter:
-    """Handles response formatting"""
-    
-    @staticmethod
-    def format_tool_response(tool_result: ToolResult) -> str:
-        """Format tool execution results into a natural response"""
-        if not tool_result.success:
-            return f"I encountered an issue: {tool_result.message}"
-        
-        result = tool_result.result
-        action = tool_result.action
-        namespace = tool_result.namespace
-        
-        if action == ProjectAction.LIST.value:
-            if result.total == 0:
-                return f"I found no projects in the '{namespace}' namespace."
-            
-            response = f"I found {result.total} project(s) in the '{namespace}' namespace:\n\n"
-            if result.projects:
-                for project in result.projects:
-                    response += f"• **{project['project_id']}**\n"
-                    response += f"  Path: `{project['path']}`\n"
-                    if project.get('description'):
-                        response += f"  Description: {project['description']}\n"
-                    response += "\n"
-            
-            return response.strip()
-        
-        elif action == ProjectAction.CREATE.value:
-            if result.success:
-                return f"✅ Successfully created project '{result.project_id}' in namespace '{namespace}'"
-            else:
-                return f"❌ Failed to create project: {result.message}"
-        
-        return str(result)
-
-    @staticmethod
-    def create_tool_info(tool_result: ToolResult) -> List[Dict]:
-        """Create tool result information for response"""
-        return [{
-            "tool_used": "projects",
-            "integration_type": tool_result.integration_type.value,
-            "action": tool_result.action,
-            "namespace": tool_result.namespace,
-            "message": f"{tool_result.integration_type.value.replace('_', ' ').title()} {'successful' if tool_result.success else 'failed'}"
-        }]
-
-class AgentFactory:
-    """Factory for creating agents with proper configuration"""
-    
-    @staticmethod
-    def create_system_prompt(capabilities: ModelCapabilities) -> SystemPromptGenerator:
-        """Create system prompt based on model capabilities"""
-        return SystemPromptGenerator(
-            background=[
-                "You are a helpful assistant for LlamaFarm project management.",
-                "You have access to a projects tool that can list and create projects in different namespaces.",
-                f"Tool calling support: {'NATIVE' if capabilities.supports_tools else 'FALLBACK'}",
-            ],
-            steps=[
-                "Analyze the user's request to determine if they need project management assistance",
-                "For listing projects: use action='list' with the appropriate namespace",
-                "For creating projects: use action='create' with namespace and project_id",
-                "Always provide clear, helpful responses based on the tool results",
-                "If tool calling fails, indicate that manual processing will be attempted",
-            ],
-            output_instructions=[
-                "Be helpful and friendly in your responses",
-                "When using tools, briefly explain what you're doing",
-                "Provide clear summaries of project operations",
-                "Use the exact namespace mentioned by the user",
-                "Format project lists in a readable way with bullet points",
-                "If you cannot use tools natively, still provide a helpful response",
-            ]
-        )
-
-    @staticmethod
-    def create_agent_config(capabilities: ModelCapabilities, system_prompt: SystemPromptGenerator, tools: List = None) -> AgentConfig:
-        """Create agent configuration"""
-        client = ModelManager.create_client(capabilities)
-        history = ChatHistory()
-        
-        config_params = {
-            "client": client,
-            "model": settings.ollama_model,
-            "history": history,
-            "system_role": "system",
-            "system_prompt_generator": system_prompt,
-            "model_api_parameters": {
-                "temperature": 0.1,
-                "top_p": 0.9,
-            }
-        }
-        
-        if capabilities.supports_tools and tools:
-            config_params["tools"] = tools
-        
-        return AgentConfig(**config_params)
-
-    @staticmethod
-    def create_agent() -> AtomicAgent[BasicChatInputSchema, BasicChatOutputSchema]:
-        """Create a new agent instance with enhanced tool integration"""
-        print(f"🔧 [Inference] Creating new agent...")
-        
-        # Get model capabilities
-        capabilities = ModelManager.get_capabilities(settings.ollama_model)
-        print(f"🔍 [Inference] Model {settings.ollama_model} tool calling support: {capabilities.supports_tools}")
-        print(f"✅ [Inference] Using {capabilities.instructor_mode.value} mode for instructor")
-
-        # Create tool instances
-        projects_tool = ProjectsTool()
-        print(f"✅ [Inference] Created ProjectsTool instance")
-        
-        # Create system prompt and agent config
-        system_prompt = AgentFactory.create_system_prompt(capabilities)
-        tools = [projects_tool] if capabilities.supports_tools else None
-        agent_config = AgentFactory.create_agent_config(capabilities, system_prompt, tools)
-        
-        if capabilities.supports_tools:
-            print(f"✅ [Inference] Added native tool support")
-        else:
-            print(f"⚠️ [Inference] No native tools added - will use manual execution")
-        
-        agent = AtomicAgent[BasicChatInputSchema, BasicChatOutputSchema](agent_config)
-        print(f"✅ [Inference] Agent created successfully")
-        return agent
-
-class ChatProcessor:
-    """Main chat processing logic"""
-    
-    @staticmethod
-    def process_chat(request: ChatRequest, session_id: str) -> Tuple[str, Optional[List[Dict]]]:
-        """Process chat request and return response with tool info"""
-        # Get or create agent
-        if session_id not in agent_sessions:
-            agent = AgentFactory.create_agent()
-            agent_sessions[session_id] = agent
-            print(f"🆕 [Inference] Created new agent session: {session_id}")
-        else:
-            agent = agent_sessions[session_id]
-            print(f"🔄 [Inference] Using existing agent session: {session_id}")
-
-        # Run agent
-        print(f"🤖 [Inference] Running agent with message: '{request.message[:100]}...'")
-        input_schema = BasicChatInputSchema(chat_message=request.message)
-        response = agent.run(input_schema)
-        
-        response_message = response.chat_message if hasattr(response, 'chat_message') else str(response)
-        print(f"📤 [Inference] Initial agent response: {response_message[:200]}...")
-        
-        # Check if manual execution is needed
-        if ResponseAnalyzer.needs_manual_execution(response_message, request.message):
-            print(f"🔧 [Inference] Template/incomplete response detected: '{response_message[:100]}...'")
-            
-            tool_result = ToolExecutor.execute_manual(request.message)
-            
-            if tool_result.success:
-                response_message = ResponseFormatter.format_tool_response(tool_result)
-                tool_info = ResponseFormatter.create_tool_info(tool_result)
-                print(f"✅ [Inference] Manual execution successful")
-            else:
-                response_message = tool_result.message
-                tool_info = ResponseFormatter.create_tool_info(tool_result)
-                print(f"❌ [Inference] Manual execution failed")
-        
-        elif MessageAnalyzer.is_project_related(request.message):
-            tool_info = [{
-                "tool_used": "projects",
-                "integration_type": IntegrationType.NATIVE.value,
-                "message": "Native tool integration used"
-            }]
-        else:
-            tool_info = None
-        
-        return response_message, tool_info
 
 # Route handlers
 @router.post("/chat", response_model=ChatResponse)
@@ -444,10 +56,7 @@ async def chat(
 @router.delete("/chat/session/{session_id}")
 async def delete_chat_session(session_id: str):
     """Delete a chat session"""
-    if session_id in agent_sessions:
-        agent_sessions[session_id].reset_history()
-        del agent_sessions[session_id]
-        print(f"🗑️ [Inference] Deleted session: {session_id}")
+    if AgentSessionManager.delete_session(session_id):
         return {"message": f"Session {session_id} deleted successfully"}
     else:
         raise HTTPException(status_code=404, detail="Session not found")
@@ -486,6 +95,8 @@ async def test_native_tools(request: ChatRequest):
     print(f"🧪 [Inference] Testing enhanced tool integration with: '{request.message}'")
     
     try:
+        from .factories import AgentFactory
+        
         agent = AgentFactory.create_agent()
         capabilities = ModelManager.get_capabilities(settings.ollama_model)
         
@@ -535,8 +146,8 @@ async def get_agent_status():
         pass
     
     return {
-        "active_sessions": len(agent_sessions),
-        "session_ids": list(agent_sessions.keys()),
+        "active_sessions": AgentSessionManager.get_session_count(),
+        "session_ids": AgentSessionManager.get_session_ids(),
         "current_model": settings.ollama_model,
         "model_supports_tools": capabilities.supports_tools,
         "integration_type": "enhanced_detection_with_fallback",
